pyproject.toml
[project]
name = "classic-models-mcp"
version = "1.0.0"
description = "MCP server for Classic Models API"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=0.9.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

